+++ packages/eslint-plugin/lib/configs/rules/possible-errors.js
@@ -1,6 +1,10 @@
 // https://eslint.org/docs/rules/#possible-errors
 
 module.exports = {
+	// Enforce “for” loop update clause moving the counter in the right direction.
+	// https://eslint.org/docs/rules/for-direction
+	'for-direction': 'error',
+
 	// Disallow calling some `Object.prototype` methods directly on objects
 	// https://eslint.org/docs/rules/no-prototype-builtins
 	'no-prototype-builtins': 'error',